requirements.txt
 
streamlit>=1.28.0
pandas>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
streamlit-aggrid>=1.0.0
cryptography>=41.0.0